% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sample_size.R
\name{get_sample_size}
\alias{get_sample_size}
\title{Determine sample size for a group sequential SMART design}
\usage{
get_sample_size(
  variances,
  beta,
  delta,
  bounds,
  n_init = 100,
  corr = bounds$corr,
  inf_frac = bounds$inf_frac,
  n_split = NULL
)
}
\arguments{
\item{variances}{A numeric vector of length \eqn{L} of variances of the
value estimators, i.e., \eqn{\sqrt{N} \times \mathrm{Cov}(\hat{\theta})}.
These should reflect the population variances rather than sample variance
or standard errors.}

\item{beta}{A numeric value between 0 and 1 specifying the type II error
rate. Power is \code{1 - beta}.}

\item{delta}{A numeric vector of length \eqn{L} of differences between
regime values and the null value (or control arm).}

\item{bounds}{A numeric vector of length \eqn{S} with the stopping
boundaries for analyses \eqn{1, \ldots, S}, or the boundaries from
\code{\link{get_bounds}} function}

\item{n_init}{A positive integer specifying the initial total trial sample
size \eqn{N} to begin the search.}

\item{corr}{A correlation matrix of dimension \eqn{L \times L} between
regime value estimators across all analyses.}

\item{inf_frac}{A numeric vector of information fractions indicating when
analyses are conducted.}

\item{n_split}{A numeric vector indicating the proportion of the sample
size at the analysis times s=1,...,S. If no argument is given, assumes
the split is proportional to the information available.}
}
\value{
A list with the following components:
\describe{
\item{N}{A numeric vector of sample sizes at each analysis.}
\item{power}{The achieved power at the final sample size.}
\item{prop_rej}{A numeric vector of cumulative rejection probabilities
at each analysis.}
\item{variances}{The input variances.}
\item{beta}{The input type II error rate.}
\item{delta}{The input alternative differences.}
\item{bounds}{The input stopping boundaries.}
\item{n_init}{The input initial sample size.}
\item{corr}{The input correlation matrix.}
\item{inf_frac}{The input information fractions.}
\item{n_split}{The input sample size split proportions.}
}
}
\description{
For specified operating characteristics, iteratively increases the sample
size until the desired power is achieved. Assumes the information fraction
remains unchanged as the sample size increases, which is reasonable for small
changes or at the design stage.
}
