% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcsttrial}
\alias{pcsttrial}
\title{pcsttrial: Simulated Clinical Trial Data for Pain Coping Skills Training}
\format{
A data frame with 284 rows and 14 variables:
\describe{
\item{pctchange}{Percent change in outcome from baseline at final assessment}
\item{height}{Patient height (cm)}
\item{weight}{Patient weight (kg)}
\item{comorbidity}{Comorbidity indicator (0 = no, 1 = yes)}
\item{painmed}{Pain medication use indicator (0 = no, 1 = yes)}
\item{chemo}{Chemotherapy indicator (0 = no, 1 = yes)}
\item{pctchangek2}{Percent change in outcome at stage 2}
\item{adherence}{Treatment adherence measure, from 0.5 to 1 in increments of 0.1}
\item{a1}{First-stage treatment assignment}
\item{a2}{Second-stage treatment assignment}
\item{r2}{Response indicator at stage 2 (0 = non-responder, 1 = responder)}
\item{study_day_enroll}{Study day of enrollment}
\item{study_day_rerand}{Study day of re-randomization}
\item{study_day_outcome}{Study day of outcome assessment}
}
}
\usage{
pcsttrial
}
\description{
A dataset containing simulated patient data to mimic the characteristics
of the Pain Coping Skills Training trial presented in Manschot, Laber, and Davidian (2023).
}
\details{
Note: this dataset is for illustrative purposes and does not use the actual trial data.
}
\keyword{datasets}
