% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rsmart-package.R
\docType{package}
\name{rsmart-package}
\alias{rsmart}
\alias{rsmart-package}
\title{rsmart: Sequential Multiple Assignment Randomized Trials design and analyses}
\description{
Implements the interim augmented inverse probability weighted estimator (IAIPWE) for estimating the value of treatment regimes embedded in sequential multiple assignment randomized trials (SMARTs). The IAIPWE subsumes both the inverse probability weighted estimator (IPWE) and the augmented inverse probability weighted estimator (AIPWE), enabling inference at interim and final analyses. The package provides functions for value estimation, sandwich variance computation, group sequential stopping boundaries, and sample size determination for multi-stage SMARTs with up to two treatment options at each stage. See Manschot, Laber, and Davidian (2023) \doi{10.1111/biom.13854} for additional details.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://MSDLLCpapers.github.io/rsmart/}
  \item \url{https://github.com/MSDLLCpapers/rsmart}
  \item \url{https://msdllcpapers.github.io/rsmart/}
  \item Report bugs at \url{https://github.com/MSDLLCpapers/rsmart/issues}
}

}
\author{
\strong{Maintainer}: Cole Manschot \email{cole.manschot@msd.com}

Other contributors:
\itemize{
  \item Tabitha Peter [contributor]
  \item Gabriela Piasecki [contributor]
  \item Laura Frederick [contributor]
}

}
\keyword{internal}
